Finishing refactor to create simple common interfaces for the main models
