+++ tests/test_gp_model.py
@@ -1,5 +1,5 @@
 import unittest
-from GCEm.gp_model import gp_model
+from GCEm import gp_model
 from GCEm.utils import get_uniform_params
 from tests.mock import *
 from numpy.testing import assert_allclose
